requirements.txt
asyncpg==0.29.0
fastapi==0.110.2
graphql-core==3.2.3
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
starlette==0.37.2
strawberry-graphql==0.227.2
uvicorn==0.29.0